+++ src/stats.hpp
@@ -4,7 +4,7 @@
 class Stats
 {
 public:
-  Stats( short startvalue = 0 )
+  Stats( int startvalue = 0 )
     : Str( startvalue ),
       Int( startvalue ),
       Wis( startvalue ),
@@ -17,14 +17,14 @@ public:
 
   }
 
-  short Str = 0;
-  short Int = 0;
-  short Wis = 0;
-  short Dex = 0;
-  short Con = 0;
-  short Cha = 0;
-  short Lck = 0;
-  short Frc = 0;
+  int Str = 0;
+  int Int = 0;
+  int Wis = 0;
+  int Dex = 0;
+  int Con = 0;
+  int Cha = 0;
+  int Lck = 0;
+  int Frc = 0;
 };
 
 #endif
